requirements.txt
Django>=4.2,<5.0
wagtail>=7.2,<8.0
wagtail-modeladmin>=2.2,<3.0
Pillow==12.0.0
reportlab==4.4.6
qrcode[pil]==8.2
django-environ==0.11.2
psycopg2-binary==2.9.9
pytest-html
mysqlclient
hijri-converter
